cbl/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.0.8
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-currencies==0.10.1
django-import-export==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-pipeline==2.0.5
environ==1.0
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
isort==5.6.4
jdcal==1.4.1
jsonfield==2.1.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
paypalrestsdk==1.13.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pyOpenSSL==19.1.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.55.0
tablib==2.0.0
toml==0.10.1
urllib3==1.26.2
virtualenv==20.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
